refactor(ui): update BubbleCard - update tests and story
